Fix `to_postfix` to properly convert infix to postfix.
while we used to properly compute the proper `Operand` order in
our convertor the `Operator` stack was incorrectly appended to the statement,
as an example:

3 + 3 * 2 was computed as:

3 3 2 + * which is obviously incorrect
3 3 2 * + is the correct postfix notation

when assembling the statement in `to_postfix` you find that the
statement is created thorughout the parsing of the given equation and
when an operator is greater than the previous one, it is placed on to a
stack. If the equation parsing is completed all elements of the stack are
moved into the statement.

in our previous example 3 + 3 * 2 would finish parsing and its state
would be as follows:

STATEMENT:  3 3 2

STACK:      | * |
            | + |
            |___|

as * has precedence over + it is expected that it would be put above it
in the stack and therefore be popped out first, so after appending one
element from the stack onto the statement you'd expect it to append *
first.

however since we were appending `stack` which was represented as
`Vec<Token>` it would take the statement into being 3 3 2 + *, to solve
this while maintaing the stack's structure you could pop each element
from the top and append it, or simply reverse the stack and append it.

NOTE: Some additional work could be done performance-wise that would
require changing the stack as the data-structure. but as of now, it is unnecessary.