Remove no longer intended TODO comments
Removed TODO comments to apply checks for an empty stack and ordering, the empty stack is irrelevant as it is pretty imperative that after emptying the stack it would be empty; and the ordering is harder to check for than it is worth. Test cases to check for that could be added.