Convert `Operand::Numeric(Numeric)` into `Operand::Numeric(i64)`
Initially `Operand::Numeric` was given the struct `Numeric` as a value
instead of i64 because I figured it would simplify further development
into Operand types that overflow the i64 value we set Numeric to be.
Evidently thinking for the future really didn't help, as it often
doesn't.

`Operand::Numeric` being given i64 instead of struct `Numeric` which
itself would hold i64 proved to be simpler and have exactly the same
result.