Fix numeral parsing error in `lexer`
This fixes an error where numerals would be lexed incorrectly as the
first digit would be incorrectly used instead of the new character. For
example:

'12' would be '1', '2' which would be read as:
'1'; would be correctly parsed and added to the operand, then...
'2'; it would here read the already converted `1` and then subtract 48 (it assumes that it is
converting a char '2' where it actually is incorrectly converting the
i64 `1` and then adding it)

You would assume '12' to be read as:
'1'; Operand = 1,
'2'; Operand = 1 + '2' -> (1 * 10) + ('2' as i64 - 48), // You would
subtract 48 to convert the ascii code to the actual value

Whereas it is actually read
'1'; Operand = 1,
'2'; Operand = 1 + 1 -> (1 * 10) + (1 as i64 - 48), // In here 1
replaces '2' incorrectly.